feat: support prepared statement parameters
This adds support for prepared statement parameters.
Parameters will be parsed from the SQL query and can be supplied using
do_put_prepared_statement_query providing a record batch with a single
row and a column for each parameter.

The QueryHandler is updated to contain the encoded parameters as a
record batch.