Add ability to set statistics in the SQLExecutor (#134)
The current `SQLExecutor` does not allow us to provide statistics, meaning that the result sets using `datafusion-federation` will always have `Rows=Absent, ...`. This can be a problem when joining large amounts of data in Datafusion, as a suboptimal plan is often selected.

This PR adds the `statistics` function to `SQLExecutor`, so it can be implemented to provide more information to the planner. It also adds it to `VirtualExecutionPlan` and `SchemaCastScanExec`, so it can reach Datafusion.

By default, it simply returns `Statistics::new_unknown`, so existing implementations of `SQLExecutor` can continue to work as normal.